app/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-json-logger
